requirements.txt
langchain
langchain-openai
langchain-community
chromadb
tiktoken
python-dotenv
langsmith
fastapi
uvicorn[standard]
pydantic